Remove misleading Matrix22::multDirMatrix
Matrix33 and Matrix44 have `operator*(VecN, MatrixNN)` that do just the
obvious mathematical N-vector X NxN-matrix.

They also define `Matrix33::multVecMatrix(Vec2 src, Vec2 dst)` and
multDirMatrix, which are homogeneous coordinate 2-vec X 3x3-mat that
assume a 3rd vector component of 1 and 0, respectively, and project
back to a 2D result with a homogeneous divide. And the same for
Matrix44::multVecMatrix and multDirMatrix, with 3-vector homogeneous
transformation.

OK, so what about Matrix22? It has `operator*(Vec2,Matrix22)` as
expected.  But it also curiously has multDirMatrix... which is wrong,
it's not a transformation on homogeneous coordinates, it's just a
regular vec*mat multiply. It is not the analogous operation to the
Matrix33 and Matrix44 methods of the same name.

So this patch removes the misnamed (and also useless) method.

Signed-off-by: Larry Gritz <lg@larrygritz.com>